requirements.txt
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
Flask==2.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.8.0
python-dotenv==0.20.0
toml==0.10.2
urllib3==1.26.9
uuid==1.30
Werkzeug==2.1.2
